requirements.txt
PyQt5>=5.15.0
openpyxl>=3.0.0
pandas>=1.0.0
python-docx>=0.8.11
